use switch

+++ install.sh
@@ -2,40 +2,40 @@
 
 if [ ! -f /usr/local/bin/pihole ]; then
 	echo "$(date) - Installing Pi-hole..."
-	curl -sSLN https://install.pi-hole.net | sudo bash
+	curl -sSL https://install.pi-hole.net | sudo bash
 fi
 
-echo "$(date) - Verifying Dependencies..."
+if [ "$1" != "un" ]; then
+	echo "$(date) - Verifying Dependencies..."
 
-PHP_VERSION=$(php -v | tac | tail -n 1 | cut -d " " -f 2 | cut -c 1-3)
-apt-get install sqlite3 $PHP_VERSION-sqlite3 jq -y
-apt-get remove speedtest-cli -y
+	apt-get remove -y speedtest-cli
+	if [ ! -f /usr/bin/speedtest ]; then
+		echo "$(date) - Adding speedtest source..."
+		# https://www.speedtest.net/apps/cli
+		curl -sSLN https://packagecloud.io/install/repositories/ookla/speedtest-cli/script.deb.sh | sudo bash
+	fi
+	PHP_VERSION=$(php -v | tac | tail -n 1 | cut -d " " -f 2 | cut -c 1-3)
+	apt-get install -y sqlite3 $PHP_VERSION-sqlite3 jq speedtest
 
-if [ ! -f /usr/bin/speedtest ]; then
-	echo "$(date) - Installing speedtest..."
-	# https://www.speedtest.net/apps/cli
-	curl -sSLN https://packagecloud.io/install/repositories/ookla/speedtest-cli/script.deb.sh | sudo bash
-	sudo apt-get install speedtest -y
+	echo "$(date) - Downloading Latest Speedtest Mod..."
+
+	cd /var/www/html
+	rm -rf new_admin
+	git clone --depth=1 https://github.com/arevindh/AdminLTE new_admin
+	cd new_admin
+	git fetch --tags -q
+	latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
+	git checkout $latestTag
+
+	cd /opt/
+	rm -rf new_pihole
+	git clone --depth=1 -b ipitio https://github.com/arevindh/pi-hole new_pihole
+	cd new_pihole
+	git fetch --tags -q
+	latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
+	git checkout $latestTag
+	chmod +x advanced/Scripts/webpage.sh
 fi
-
-echo "$(date) - Downloading Latest Speedtest Mod..."
-
-cd /var/www/html
-rm -rf new_admin
-git clone --depth=1 https://github.com/arevindh/AdminLTE new_admin
-cd new_admin
-git fetch --tags -q
-latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
-git checkout $latestTag
-
-cd /opt/
-rm -rf new_pihole
-git clone --depth=1 https://github.com/arevindh/pi-hole new_pihole
-cd new_pihole
-git fetch --tags -q
-latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
-git checkout $latestTag
-chmod +x advanced/Scripts/webpage.sh
 
 db=$([ "$1" == "up" ] && echo "$3" || [ "$1" == "un" ] && echo "$2" || echo "$1")
 curl -sSLN https://github.com/arevindh/pihole-speedtest/raw/master/uninstall.sh | sudo bash -s -- $db
@@ -47,6 +47,9 @@ fi
 
 if [ "$1" == "up" ]; then
 	echo "$(date) - Updating Pi-hole..."
+	cd /var/www/html/admin
+	git reset --hard origin/master
+	git switch -f master
 	PIHOLE_SKIP_OS_CHECK=true sudo -E pihole -up
 	if [ "$2" == "un" ]; then
 		rm -rf /opt/pihole/webpage.sh.*
@@ -66,7 +69,7 @@ rm -rf org_admin
 mv admin org_admin
 cp -r new_admin mod_admin
 mv new_admin admin
-cd -
+cd - > /dev/null
 cp pihole/webpage.sh.mod pihole/webpage.sh
 
 if [ ! -f /etc/pihole/speedtest.db ] || [ "$db" == "db" ]; then
